Fix inconsitencies with openinverter_can_tool can mapping (#7)
* Don't try to erroneously setNoDelay on the WebServer

During setup() the code attempts to disable the Nagle
algorithm on a non-existent client socket which causes
an ugly error in the system log:

[   432][E][WiFiClient.cpp:320] setSocketOption(): fail on -1, errno: 9, "Bad file number"

Remove this code. If setNoDelay() needs to be set then
this needs to be done on a per-connection basis.

* Fix can mapping with negative gains and offsets

When reading can mappings the 24-bit gain value was being misinterpreted. Fix this by sign-extending the 24-bit value into a regular int32_t.

When creating a can mapping the offset field was being corrupted. Correctly mask off the gain before ORing the shifted offset value.

Tests:
 - Add can mappings with max positive and negative
    values for gain and offset.
 - Verify list displays the expected map
 - Verify with OIC that the map is as expected
 - Load a complex can map with a mix of big and little
    endian mappings with OIC and verify the
    display matches